had to switch to iOS only as my only machine is Apple Silicon and the Android Emulator failed working after a short while. Completed upgrading system, have to fix lucky donations and improve on CSS

+++ ZooClicker/Models/FrogHabitat.cs
@@ -3,7 +3,7 @@ namespace ZooClicker.Models
 {
     public class FrogHabitat : Habitat
     {
-        public FrogHabitat(int lvl = 1, float cst = 100, float don = 10, string n = "Frog")
+        public FrogHabitat(int lvl = 1, int cst = 100, int don = 10, string n = "Frog")
             : base(lvl, cst, don, n)
         {
         }
